refactor(server): extract workspace lifecycle actions
