+++ server/proliferate/server/cloud/workspaces/lifecycle/models.py
@@ -1,0 +1,7 @@
+from __future__ import annotations
+
+from pydantic import BaseModel
+
+
+class WorkspaceLifecycleMutationResponse(BaseModel):
+    ok: bool = True
